backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cohere==5.5.3